1.新增了任务计划工具 2.支持以命令行启动RPC服务器并支持按配置启动提供服务 3.实现了守护任务管理器以自动条件式启动守护任务 4.支持了连接池实例断开后自动重连 5.支持了多种组件日志及可配置是否开启（后续还会升级） 6.新增了批量任务按步迭代处理器 7.SwooleUtility、Structure、Utility等工具类新增了一些工具方法 8.连接建立事件支持转发到控制器处理 9.修复了分库中间件并发请求时异常可能捕获不到的问题 10.修复了原始语句构建器绑定的参数可能丢失的问题 11.解决了以命令行执行完协程任务后无法立即退出的问题 12.解决了Session::renew()后服务端异常无法将新sid传到客户端导致客户端会话失效的问题 13.调整并解决了SessionManager逻辑Bug导致会话丢失的问题 14.优化了大量细节代码及Bug修复（自行通过提交记录查看） 15.本次提交了大量不兼容旧版的更新，请慎重。且待PHP正式发布8.1后，还会进行大量不兼容更新。

+++ engine/project/session/SessionRedis.php
@@ -6,11 +6,13 @@
 
 namespace dce\project\session;
 
+use dce\project\request\RequestManager;
 use dce\storage\redis\DceRedis;
 use Redis;
 
 class SessionRedis extends Session {
-    private const META_KEY = '&meta';
+    private const MetaKey = '&meta';
+    private const ReferenceKey = 'reference';
 
     /**
      * 更新session过期时间
@@ -19,8 +21,19 @@ class SessionRedis extends Session {
     protected function touch(mixed $redis = null): void {
         $rKey = $this->getId(true);
         // 如果是长存时间，则session需续长存时间，否则需续短存时间
-        $ttl = ($redis->hGet($rKey, self::META_KEY)['long_live'] ?? 0) && self::$config['long_ttl'] ? self::$config['long_ttl'] : (self::$config['ttl'] ?: 3600);
+        $meta = $redis->hGet($rKey, self::MetaKey);
+        $ttl = ($meta['long_live'] ?? 0) && self::$config['long_ttl'] ? self::$config['long_ttl'] : (self::$config['ttl'] ?: 3600);
         $redis->expire($rKey, $ttl);
+        if (
+            key_exists(self::ReferenceKey, (array) $meta)
+            && ([$requestId, $referenceSid] = $meta[self::ReferenceKey])
+            && $requestId !== RequestManager::currentId()
+            && $referenceSid !== $this->getId()
+        ) {
+            unset($meta[self::ReferenceKey]);
+            $redis->hSet($rKey, self::MetaKey, $meta);
+            self::newBySid($referenceSid)->destroy();
+        }
     }
 
     /** @inheritDoc */
@@ -61,7 +74,6 @@ class SessionRedis extends Session {
     public function delete(string $key): void {
         $redis = DceRedis::get(self::$config['index']);
         $redis->hDel($this->getId(true), $key);
-        $this->tryTouch($redis);
         DceRedis::put($redis);
     }
 
@@ -74,14 +86,30 @@ class SessionRedis extends Session {
 
     /** @inheritDoc */
     public function renew(bool $longLive = false): static {
-        $data = $this->getAll();
-        $data[self::META_KEY] = ['create_time' => time(), 'long_live' => $longLive];
-        $this->destroy();
-        $this->setId(self::genId());
+        $newSid = self::genId();
+        $requestId = RequestManager::currentId();
         $redis = DceRedis::get(self::$config['index']);
+        $data = $redis->hGetAll($this->getId(true));
+
+        if ($data) {
+            if ($requestId) {
+                $oldMeta = $data[self::MetaKey];
+                $oldMeta[self::ReferenceKey] = [$requestId, $newSid];
+                $redis->hSet($oldKey = $this->getId(true), self::MetaKey, $oldMeta);
+                $redis->expire($oldKey, self::$oldTtl);
+            } else {
+                $this->destroy();
+            }
+        }
+
+        $data[self::MetaKey] = ['create_time' => time(), 'long_live' => $longLive];
+        $requestId && $data && $data[self::MetaKey][self::ReferenceKey] = [$requestId, $this->getId()];
+
+        $this->setId($newSid);
         $redis->hMSet($this->getId(true), $data);
         $this->touched = false;
         $this->tryTouch($redis);
+
         DceRedis::put($redis);
         return $this;
     }
@@ -90,7 +118,7 @@ class SessionRedis extends Session {
     public function getMeta(): array {
         $redis = DceRedis::get(self::$config['index']);
         $rKey = $this->getId(true);
-        $meta = $redis->hGet($rKey, self::META_KEY) ?: [];
+        $meta = $redis->hGet($rKey, self::MetaKey) ?: [];
         $meta += ['expiry' => ($meta['long_live'] ?? 0) ? self::$config['long_ttl'] : self::$config['ttl'], 'ttl' => $redis->ttl($rKey)];
         DceRedis::put($redis);
         return $meta;
